move code to .cpp files
